refactored search Input
